fix(planning): make /planning redirect to the forecast rather than be it
Planning is a heading over two things, not a page in its own right. The forecast
sat at /planning itself, which left the secondary navigation quietly dishonest:
the Forecast item pointed at the section root rather than at a URL that is the
forecast, so it could not distinguish "on the forecast" from "somewhere under
Planning". The forecast now lives at /planning/forecast and the bare path
redirects to it, following the same pattern /settings already uses.

The dashboard widget links straight to /planning/forecast rather than through the
redirect. Router loaders and beforeLoad redirects have deadlocked in-app
navigation here before, so drill-downs go directly to their destination; the
redirect stays as the bookmark and typed-URL fallback, which is what it is for.

Verified in the browser: a direct load of /planning lands on the forecast, and
so does an in-app click on the Planning breadcrumb from the retirement page —
the case that would deadlock if a loader were ever added above these routes.
Both nav items highlight correctly, which was the point.
